requirements.txt
python-telegram-bot>=20.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pytz>=2024.1
